# api/openapi.policy.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/me/account" : {
      "delete" : {
        "operationId" : "withdrawMember",
        "parameters" : [ {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "탈퇴 완료"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "503" : {
            "description" : "저장소 오류로 탈퇴 미완료"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "회원 탈퇴와 개인 데이터·모든 로그인 세션 삭제",
        "tags" : [ "member-controller" ]
      }
    },
    "/api/v1/me/conditions" : {
      "delete" : {
        "operationId" : "clearMemberConditions",
        "parameters" : [ {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 저장 조건 삭제",
        "tags" : [ "member-controller" ]
      },
      "get" : {
        "operationId" : "getMemberConditions",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MemberConditions"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내가 저장한 기본 조건 조회",
        "tags" : [ "member-controller" ]
      },
      "put" : {
        "operationId" : "saveMemberConditions",
        "parameters" : [ {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BasicConditions"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "확인한 내 기본 조건 저장",
        "tags" : [ "member-controller" ]
      }
    },
...
    "/api/v1/me/notifications" : {
      "get" : {
        "description" : "최신순으로 정렬하고 필터를 전체 알림에 적용한 뒤 페이지를 나눈다. unreadCount는 필터와 무관한 회원 전체의 안 읽은 알림 수다.",
        "operationId" : "listMemberNotifications",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "filter",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "ALL",
            "enum" : [ "ALL", "UNREAD" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MemberNotificationList"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 알림 페이지·안 읽은 알림 수 조회",
        "tags" : [ "member-controller" ]
      }
    },
    "/api/v1/me/notifications/read-all" : {
      "post" : {
        "description" : "페이지·필터와 관계없이 본인의 미읽음 알림 전체를 처리한다. 갱신 쿼리 시작 후 도착한 알림과 기존 읽은 시각은 유지한다.",
        "operationId" : "readAllMemberNotifications",
        "parameters" : [ {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 알림 모두 읽음 처리",
        "tags" : [ "member-controller" ]
      }
    },
    "/api/v1/me/notifications/{id}/read" : {
      "post" : {
        "operationId" : "readMemberNotification",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 알림 읽음 처리",
        "tags" : [ "member-controller" ]
      }
    },
    "/api/v1/me/policies" : {
      "get" : {
        "operationId" : "listSavedPolicies",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SavedPolicyList"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 관심 정책과 최신 마감 일정 조회",
        "tags" : [ "member-controller" ]
      }
    },
    "/api/v1/me/policies/{number}" : {
      "delete" : {
        "operationId" : "removeSavedPolicy",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 관심 정책 해제와 미발송 알림 취소",
        "tags" : [ "member-controller" ]
      },
      "put" : {
        "operationId" : "savePolicy",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "내 관심 정책 저장과 마감 알림 예약",
        "tags" : [ "member-controller" ]
      }
    },
    "/api/v1/me/policies/{number}/changes" : {
      "get" : {
        "description" : "본인이 저장한 정책만 조회한다. 같은 정책번호의 저장 당시 개정과 현재 공개 개정을 비교하며 저장 기준이나 알림 상태를 변경하지 않는다.",
        "operationId" : "getSavedPolicyChanges",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SavedPolicyChanges"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관심 정책의 저장 당시 내용과 현재 내용 조회",
        "tags" : [ "member-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BasicConditions" : {
        "type" : "object",
        "properties" : {
          "birthDate" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          },
          "district" : {
            "type" : [ "string", "null" ],
            "enum" : [ "강남구", "강동구", "강북구", "강서구", "관악구", "광진구", "구로구", "금천구", "노원구", "도봉구", "동대문구", "동작구", "마포구", "서대문구", "서초구", "성동구", "성북구", "송파구", "양천구", "영등포구", "용산구", "은평구", "종로구", "중구", "중랑구", null ]
          },
          "employmentStatus" : {
            "type" : [ "string", "null" ],
            "enum" : [ "EMPLOYED", "SELF_EMPLOYED", "NOT_EMPLOYED", "FREELANCER", "DAY_WORKER", "ENTREPRENEUR", "SHORT_TERM_WORKER", "FARMER", "OTHER", null ]
          }
        }
      },
...
      "MemberConditions" : {
        "type" : "object",
        "properties" : {
          "conditions" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/BasicConditions"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "conditions" ]
      },
...
      "MemberNotification" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "read" : {
            "type" : "boolean"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "id", "message", "policyNumber", "read", "title" ]
      },
      "MemberNotificationList" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MemberNotification"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unreadCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "hasNext", "items", "page", "pageSize", "total", "unreadCount" ]
      },
...
      "PolicyApiError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ]
      },
...
      "PolicyContent" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyOfficialLink"
            }
          },
          "organization" : {
            "type" : "string"
          },
          "regionCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "sections" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyTextSection"
            }
          },
          "sourceModifiedAtText" : {
            "type" : "string",
            "description" : "시간대가 확정되지 않은 원천 수정 일시 원문"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "category", "description", "links", "organization", "regionCodes", "sections", "sourceModifiedAtText", "title" ]
      },
...
      "PolicyDeadline" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          },
          "note" : {
            "type" : "string"
          }
        },
        "required" : [ "date", "note" ]
      },
...
      "PolicyOfficialLink" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "required" : [ "label", "url" ]
      },
...
      "PolicyRecruitment" : {
        "type" : "object",
        "properties" : {
          "daysUntilDeadline" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "평가 시점의 서울 날짜부터 마감일까지 남은 일수. 마감일이 없으면 null"
          },
          "deadlineOnSeoul" : {
            "type" : [ "string", "null" ],
            "format" : "date",
            "description" : "확인한 마감일(서울). 상시·소진형·기간 미확인과 날짜 없는 마감은 null"
          },
          "evaluatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "explanation" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "BEFORE_OPENING", "OPEN", "CLOSED", "ROLLING", "UNTIL_EXHAUSTED", "UNKNOWN" ]
          }
        },
        "required" : [ "daysUntilDeadline", "deadlineOnSeoul", "evaluatedAt", "explanation", "status" ]
      },
...
      "PolicyTextSection" : {
        "type" : "object",
        "properties" : {
          "text" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "text", "title" ]
      },
      "SavedPolicy" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "currentRevision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deadline" : {
            "$ref" : "#/components/schemas/PolicyDeadline"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "recruitment" : {
            "$ref" : "#/components/schemas/PolicyRecruitment"
          },
          "savedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "savedRevision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "currentRevision", "deadline", "policyNumber", "recruitment", "savedAt", "savedRevision", "title" ]
      },
      "SavedPolicyChanges" : {
        "type" : "object",
        "properties" : {
          "current" : {
            "$ref" : "#/components/schemas/SavedPolicyVersion"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "saved" : {
            "$ref" : "#/components/schemas/SavedPolicyVersion"
          },
          "savedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "current", "policyNumber", "saved", "savedAt" ]
      },
      "SavedPolicyList" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SavedPolicy"
            }
          }
        },
        "required" : [ "items" ]
      },
      "SavedPolicyVersion" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sourceCapturedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "content", "revision", "sourceCapturedAt" ]
      }
    },
    "securitySchemes" : {
      "memberSession" : {
        "description" : "카카오·네이버 로그인 후 발급하는 서버 세션",
        "in" : "cookie",
        "name" : "YPM_SESSION",
        "type" : "apiKey"
      }
    }
  }
}